v1/requirements.txt
pyyaml~=5.4.1
requests~=2.26.0